Add all the bootstraps.
Got the first test running while porting a large amount of code.  The
next set of commits begins the more test driven nature of development
but wanted to get the main module ported and workable (which involved
porting some other modules).

There is one test currently of the kernel_index function in
upkern.sources but more tests are behind it.